@@ -1,16 +1,15 @@
 import React, { Component } from "react";
-import Talambanreport from "./Talambanreport/Talambanreport";
+import RefillingReport from "./MakeReports/RefillingReport";
 import Formexample from "./Formexample/Formexample";
 import Roadtoreact from "./Roadtoreact/Roadtoreact";
 import Todos from "./Todos/Todos";
 import Tabsreact from "./Tabsreact/Tabsreact";
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  Link,
-  Redirect,
-} from "react-router-dom";
+import Moneymain from "./MoneyCounter/Moneymain";
+import Refillingcc from "./Monitorcctv/Refillingcc";
+
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+
+
 
 export default class HomePage extends Component {
   constructor(props) {
@@ -20,16 +19,18 @@ export default class HomePage extends Component {
   render() {
     return (
       <Router>
-        <Switch>
+        <Routes>
           <Route exact path="/">
             Please use another url
           </Route>
-          <Route path="/talamban" component={Talambanreport} />
-          <Route path="/formex" component={Formexample} />
-          <Route path="/todos" component={Todos} />
-          <Route path="/rtr" component={Roadtoreact} />
-          <Route path="/rtab" component={Tabsreact} />
-        </Switch>
+          <Route path="/refillingreport" element={<RefillingReport />} />
+          <Route path="/formex" element={<Formexample />} />
+          <Route path="/todos" element={<Todos />} />
+          <Route path="/rtr" element={ <Roadtoreact />} />
+          <Route path="/rtab" element={ <Tabsreact />} />
+          <Route path="/moncount" element={<Moneymain />} />
+          <Route path="/cctv/talamban-model" element={<Refillingcc />} /> 
+        </Routes>
       </Router>
     );
   }
