@@ -1,13 +1,12 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
-import 'bootstrap/dist/css/bootstrap.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('arsoreact')
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+    <App />
 );
 
 // If you want to start measuring performance in your app, pass a function
