@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import path, include
 from .views import index, free
 
 urlpatterns = [
@@ -6,10 +6,13 @@ urlpatterns = [
     path('join', index),
     path('create', index),
     path('labangon', free),
-    path('talamban',free),
+    path('talamban', free),
     path('formex', index),
-    path('todos',index),
-    path('rtr',index),
-    path('join/1', index),
-    path('rtab',index)
+    path('todos', index),
+    path('moncount', index),
+    path('rtr', index),
+    path(r'join/1', index),
+    path('', include('busyCctvCounter.urls')),
+    path(r'cctv/talamban-model/', free),
+    path('rtab', index)
 ]
